refactor: 카드 목록 화면 관련 패키지 수정
