+++ app/src/main/java/nextstep/payments/MainActivity.kt
@@ -1,6 +1,5 @@
 package nextstep.payments
 
-import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -8,8 +7,8 @@ import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.compose.setContent
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
-import nextstep.payments.screen.CardListScreen
-import nextstep.payments.screen.CardListViewModel
+import nextstep.payments.screen.cardlist.CardListScreen
+import nextstep.payments.screen.cardlist.CardListViewModel
 import nextstep.payments.screen.newcard.NewCardActivity
 import nextstep.payments.ui.theme.PaymentsTheme
